requirements.txt
awscli==1.14.58
boto3==1.6.11
botocore==1.9.11
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
cookies==2.2.1
cycler==0.10.0
docutils==0.14
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
jmespath==0.9.3
kiwisolver==1.0.1
lxml==4.1.1
matplotlib==2.2.0
matplotlib-venn==0.11.5
mlbgame==2.4.2
numpy==1.14.1
oauth2client==4.1.2
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.8.2
python-forecastio==1.3.5
pytz==2018.3
PyYAML==3.12
requests==2.18.4
responses==0.8.1
rsa==3.4.2
s3transfer==0.1.13
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.5
uritemplate==3.0.0
urllib3==1.22
